bug correction when installer is generated on Windows OS
Important: requirement for matplotlib is 3.2.2 otherwise data collection is not correct (pyinstaller add venv, tests, test_data, ... files).
Pyinstaller is not working with matplotlib greater than 3.2.2

+++ tasks.py
@@ -1,8 +1,13 @@
 # -*- coding: utf-8 -*-
 import sys
+import os
 from pathlib import Path
 
 from invoke import task
+
+
+__ABSPATH__ = Path(__file__).parent
+__version__ = "v2021.1.0a1"
 
 
 @task
@@ -15,17 +20,16 @@ def clean(c, docs=False, bytecode=False, extra=''):
     if extra:
         patterns.append(extra)
     for pattern in patterns:
-        c.run("rm -rf {}".format(pattern))
+        print(pattern)
+        if sys.platform in ('win32', 'cygwin'):
+            abs_path = __ABSPATH__.joinpath(pattern)
+            if os.path.exists(abs_path):
+                c.run("@RD /S /Q {}".format(abs_path))
+        else:
+            c.run("rm -rf {}".format(pattern))
 
 
-@task(clean)
-def build(c, docs=False):
-    c.run("python setup.py build")
-    if docs:
-        c.run("sphinx-build docs docs/_build")
-
-
-@task
+@task(pre=[clean])
 def freeze(c):
     app_name = r'PlottingApp'
     icon_path = rf'{Path(__file__).parent.joinpath("resources/icons/PlottingApp.ico")}'
@@ -47,9 +51,8 @@ def freeze(c):
 @task(pre=[freeze])
 def installer(c):
     if sys.platform in ('win32', 'cygwin'):
-        # c.run(f"makensis /V4 /DVERSION={__version__} windows.nsi")
-        c.run("cd resources/installer")
-        c.run("makensis /V4 windows.nsi")
+        nsi_file = __ABSPATH__.joinpath("resources/installer/windows.nsi")
+        c.run(f'makensis /V4 /DVERSION="{__version__}" "{nsi_file}"')
     elif sys.platform in ('darwin', 'linux'):
         pass
     else:
